requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# LangChain and LLM Integration
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0

# Dataset Loading
datasets>=2.0.0
huggingface-hub>=0.16.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.10.0

# Optional: For local model inference
# torch>=1.11.0
# transformers>=4.20.0

# Development and Testing
pytest>=6.0.0
jupyter>=1.0.0